creacion del registrarse y cockies con sid funciona todo al final de la clase del 10/11

+++ blog/borrar_mensaje.php
@@ -1,7 +1,10 @@
 <?php
+    session_start();
+
     require_once "modelos/connexionDB.php";
     require_once "modelos/mensaje.php";
     require_once "modelos/mensajesDAO.php";
+    require_once 'funciones.php';
 
     //Creamos la conexión usando la clase que hemos creado
     $connexionDB = new connexionDB('root','','localhost','blog');
@@ -12,8 +15,15 @@
 
     //Obtener el mensaje
     $idMensaje = htmlspecialchars($_GET['id']);
-
-    $mensajesDAO->delete($idMensaje);
+    $mensaje = $mensajesDAO->getById($idMensaje);
+    
+    //Comprobamos que mensaje pertenece al usuario conectado
+    if($_SESSION['id']== $mensaje->getIdUsuario()){
+        $mensajesDAO->delete($idMensaje);
+    }else{
+        $guardarMensaje("No puedes borrar este mensaje");
+    }
+  
 
     header('location: index.php');
